pyproject.toml
[tool.poetry]
name = "dspt-ui-automation-playwright-bdd"
version = "0.1.0"
description = "UI automation framework for DSPT project using Playwright + Behave"
authors = ["Sneha Kawale <drivesneha123@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
playwright = "^1.48.0"
behave = "^1.2.6"
allure-python-commons = "^2.13.5"
pytest = "^8.3.3"
requests = "^2.32.3"
faker = "^30.8.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^4.3.0"
allure-pytest = "^2.15.0"
allure-behave = "^2.15.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
